backend/requirements.txt
# Core Django
Django==4.2.25
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Production Server
gunicorn==21.2.0

# Image Processing
# Note: For Python 3.14, consider using Python 3.11 or 3.12 instead
# If you must use 3.14, try: pip install opencv-python-headless
opencv-python-headless
Pillow==12.0.0

# HTTP & Utilities
requests==2.32.5
sqlparse==0.5.3
